% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historic_prop_pul_tb.R
\name{plot_historic_prop_pul_tb}
\alias{plot_historic_prop_pul_tb}
\title{Plot Historic Proportions of Pulmonary/Extra-Pulmonary TB}
\usage{
plot_historic_prop_pul_tb(df = tb_not_ew, plot_theme = NULL,
  colour_scale = NULL, return = FALSE)
}
\arguments{
\item{df}{A dataframe, (defaults to using \code{tb_not_ew})
of historic TB notifications}

\item{plot_theme}{The ggplot2 theme to use, defaults to \code{\link[ggplot2]{theme_minimal}}.}

\item{colour_scale}{The colour scale to plot with, defaults to \code{\link[ggplot2]{scale_colour_viridis_d}}}

\item{return}{A logical indiciting if the plot should be printed or returned (defaults to \code{FALSE})}
}
\value{
A ggplot2 plot of TB notifications over time, with a secondary zoomed plot from a specified date.
}
\description{
Exported from \href{https://www.samabbott.co.uk/tbinenglanddataclean/}{tbinenglanddataclean}.
}
\examples{

plot_historic_prop_pul_tb()
}
