% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historic_tb_ew.R
\name{plot_historic_tb_ew}
\alias{plot_historic_tb_ew}
\title{Plot Historic TB Notifications in England and Wales}
\usage{
plot_historic_tb_ew(df = biddmodellingcourse::tb_not_ew,
  df_interventions = biddmodellingcourse::tb_interventions_timeline,
  zoom_date_start = 1982, zoom_date_end = NULL, plot_theme = NULL,
  colour_scale = NULL, return = FALSE)
}
\arguments{
\item{df}{A dataframe, (defaults to using \code{tb_not_ew})
of historic TB notifications}

\item{df_interventions}{A dataframe, (defaults to using \code{tb_interventions_timeline})
of TB interventions, focussing on there history in the UK. If passed as \code{NULL} no interventions will be plotted.}

\item{zoom_date_start}{Numeric, the year to filter notifications from for the second zoomed plot.}

\item{zoom_date_end}{Numeric, the year to end the zoom window, defaults to the final year of the data.}

\item{plot_theme}{The ggplot2 theme to use, defaults to \code{\link[ggplot2]{theme_minimal}}.}

\item{colour_scale}{The colour scale to plot with, defaults to \code{\link[ggplot2]{scale_colour_viridis_d}}}

\item{return}{A logical indiciting if the plot should be printed or returned (defaults to \code{FALSE})}
}
\value{
A ggplot2 plot of TB notifications over time, with a secondary zoomed plot from a specified date.
}
\description{
Exported from \href{https://www.samabbott.co.uk/tbinenglanddataclean/}{tbinenglanddataclean}.
}
\examples{
plot_historic_tb_ew()
}
