% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot Compartment Populations over Time for a Model Simulation}
\usage{
plot_model(sim, facet = TRUE, interactive = FALSE)
}
\arguments{
\item{sim}{A tibble of model output as formated by \code{\link[biddmodellingcourse]{solve_ode}}}

\item{facet}{Logical, defaults to \code{TRUE}. If \code{FALSE} then the plot will not be facetted
otherwise it will be.}

\item{interactive}{Logical, defaults to \code{FALSE}. If \code{TRUE} produces an interative plot.}
}
\value{
A Plot of each model compartments population over time.
}
\description{
Make seperate plots for each model compartment. Assumes model output is structured
as that produced from \code{\link[biddmodellingcourse]{solve_ode}}.
}
\examples{

## Intialise
N = 100000
I_0 = 1
S_0 = N - I_0
R_0 = 1.1
beta = R_0

##Time for model to run over
tbegin = 0
tend = 50
times <- seq(tbegin, tend, 1)

##Vectorise input
parameters <- as.matrix(c(beta = beta))
inits <- as.matrix(c(S = S_0, I = I_0))

sim <- solve_ode(model = SI_ode, inits, parameters, times, as.data.frame = TRUE)

plot_model(sim, facet = FALSE, interactive = FALSE)

plot_model(sim, facet = TRUE, interactive = FALSE)
}
