% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_table.R
\name{pretty_table}
\alias{pretty_table}
\title{Pretty tables}
\value{
A markdown table
}
\description{
Imported from \href{https://www.samabbott.co.uk/prettypublisher/}{prettypublisher}.
}
\examples{

}
\keyword{internal}
