% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_interventions_timeline.R
\docType{data}
\name{tb_interventions_timeline}
\alias{tb_interventions_timeline}
\title{Tuberculosis Interventions Timeline}
\format{A data frame with 11 rows and 5 variables.
\describe{
\item{year}{Calendar year}
\item{intervention}{The name of the intervention}
\item{type}{Type of intervetion, such as antibitoic, vaccine or strategy}
\item{line}{For antibiotics the order of precedance in which it is used}
\item{detail}{Detail on the interventio}
}}
\usage{
tb_interventions_timeline
}
\description{
A dataset containing a time of Tuberculosis interventions. Aggregated from multiple sources
but based in part \href{https://en.wikipedia.org/wiki/Timeline_of_tuberculosis}{this timeline}.
}
\keyword{datasets}
